+++ packages/gen-typescript-declarations/src/test/goldens/custom/my-element.d.ts
@@ -15,7 +15,7 @@
  *
  *   /** More danger! *\/*\/
  */
-interface MyElementElement extends Polymer.Element {
+interface MyElementElement extends Polymer.LegacyElementMixin, HTMLElement {
 }
 
 interface HTMLElementTagNameMap {