+++ packages/gen-typescript-declarations/src/test/goldens/polymer-3/lib/legacy/class.d.ts
@@ -27,26 +27,6 @@ export {mixinBehaviors};
  */
 declare function mixinBehaviors<T>(behaviors: object|object[], klass: {new(): T}): {new(): T};
 
-declare class PolymerGenerated {
-  created(): void;
-  _registered(): void;
-  _applyListeners(): void;
-  _ensureAttributes(): void;
-  ready(): void;
-  attached(): void;
-  detached(): void;
-
-  /**
-   * Implements native Custom Elements `attributeChangedCallback` to
-   * set an attribute value to a property via `_attributeToProperty`.
-   *
-   * @param name Name of attribute that changed
-   * @param old Old attribute value
-   * @param value New attribute value
-   */
-  attributeChanged(name: string, old: string|null, value: string|null): void;
-}
-
 export {Class};