+++ packages/analyzer/CHANGELOG.md
@@ -9,6 +9,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 * The `Function` type is now exported from `index.js`.
 * `Import` features now have an `astNodePath`.
 * `WarningStringifyOptions` now takes an optional `maxCodeLines` to print.
+* Elements registered with the legacy `Polymer()` function now have
+  `isLegacyPolymerFactoryCall: true`.
+* `@extends {SuperClass}` annotations are now parsed. Previously the type was
+  not extracted if it had curly braces.
+* Expressions annotated with `@constructor` will now be scanned as classes.
+* Ephemeral super classes will now be collapsed into their child classes.
 <!-- Add new, unreleased changes here. -->
 
 ## [3.0.2] - 2018-06-28