+++ packages/gen-typescript-declarations/CHANGELOG.md
@@ -12,6 +12,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Automatically detect if a project uses NPM or Bower and configure module
   resolution settings accordingly.
 - Automatically import/export synthetic mixin constructor interfaces.
+- Superclasses and mixins are now emitted for classes.
+- Element super classes are now emitted.
+- Legacy Polymer elements now extend `LegacyElementMixin` and `HTMLElement`
+  instead of `PolymerElement`.
+- Mixin instance interfaces now extend the instance interfaces for the mixins
+  that they automatically apply.
 
 ## [1.3.0] - 2018-06-29
 - Generate typings for class constructors.